added tests

+++ routes/tags.js
@@ -12,8 +12,8 @@ router.use('/', passport.authenticate('jwt', { session: false, failWithError: tr
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
-
-  Tag.find()
+  const userId = req.user.id;
+  Tag.find({userId})
     .sort('name')
     .then(results => {
       res.json(results);
@@ -26,7 +26,7 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
-
+  const userId = req.user.id;
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -34,7 +34,7 @@ router.get('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Tag.findById(id)
+  Tag.findOne({_id:id, userId})
     .then(result => {
       if (result) {
         res.json(result);
@@ -50,8 +50,8 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { name } = req.body;
-
-  const newTag = { name };
+  const userId = req.user.id;
+  const newTag = { name, userId};
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -77,7 +77,7 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const { id } = req.params;
   const { name } = req.body;
-
+  const userId = req.user.id;
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -91,9 +91,9 @@ router.put('/:id', (req, res, next) => {
     return next(err);
   }
 
-  const updateTag = { name };
+  const updateTag = { name, userId};
 
-  Tag.findByIdAndUpdate(id, updateTag, { new: true })
+  Tag.findByIdAndUpdate({_id:id, userId}, updateTag, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -113,7 +113,7 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
-
+  const userId = req.user.id;
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -121,7 +121,7 @@ router.delete('/:id', (req, res, next) => {
     return next(err);
   }
 
-  const tagRemovePromise = Tag.findByIdAndRemove(id);
+  const tagRemovePromise = Tag.findByIdAndRemove({_id:id, userId});
 
   const noteUpdatePromise = Note.updateMany(
     { tags: id },
